\name{datasetpgnmp1}
\alias{datasetpgnmp1}
\docType{data}
\title{Dataset 1 of the Monty Python method}
\description{The dataset contains tail algorithm constants for sampling from the tail of the \eqn{p}-generalized normal distribution in context of a simulation of the \eqn{p}-generalized normal distribution with the Monty Python method.}
\usage{data(datasetpgnmp1)}


\examples{data(datasetpgnmp1)}
\keyword{datasets}
