\name{datasetpgnmp2}
\alias{datasetpgnmp2}

\docType{data}
\title{Dataset 2 of the Monty Python method}
\description{The dataset contains optimal rectangle widths in context of a simulation of the \eqn{p}-generalized normal distribution with the Monty Python method.}
\usage{data(datasetpgnmp2)}


\examples{data(datasetpgnmp2)}
\keyword{datasets}
