\name{datasetpgnzig}
\alias{datasetpgnzig}

\docType{data}
\title{Dataset of the Ziggurat method}
\description{The dataset contains tail algorithm constants for sampling from the tail of the \eqn{p}-generalized normal distribution in context of a simulation of the \eqn{p}-generalized normal distribution with the Ziggurat method.}
\usage{data(datasetpgnzig)}


\examples{data(datasetpgnzig)}
\keyword{datasets}
